requirements.txt
aiohttp==3.6.2
appnope==0.1.0
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.2
django-extensions==3.0.9
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.10
ipython-genutils==0.2.0
jedi==0.17.2
multidict==4.7.6
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
psycopg2==2.8.6
ptyprocess==0.6.0
Pygments==2.7.1
pytz==2020.1
requests==2.24.0
slackclient==2.9.1
sqlparse==0.3.1
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
whitenoise==5.2.0
yarl==1.6.0
